pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "travel-map"
version = "0.1.0"
description = "Generate beautiful travel maps from YAML config with multiple visual styles"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "folium>=0.15.0",
    "matplotlib>=3.8.0",
    "cartopy>=0.22.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
    "click>=8.1.0",
    "numpy>=1.24.0",
]

[project.scripts]
travel-map = "travel_map.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
